+++ app/lib/shared/versions.dart
@@ -21,9 +21,9 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
-  '2021.06.08', // The current [runtimeVersion].
+  '2021.06.14', // The current [runtimeVersion].
+  '2021.06.08',
   '2021.06.01',
-  '2021.05.25',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -53,10 +53,10 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.12.0';
-final String toolStableDartSdkVersion = '2.13.1';
-final String toolStableFlutterSdkVersion = '2.2.1';
-final String toolPreviewDartSdkVersion = '2.13.1';
-final String toolPreviewFlutterSdkVersion = '2.2.1';
+final String toolStableDartSdkVersion = '2.13.3';
+final String toolStableFlutterSdkVersion = '2.2.2';
+final String toolPreviewDartSdkVersion = '2.13.3';
+final String toolPreviewFlutterSdkVersion = '2.2.2';
 
 // Value comes from package:pana.
 final String panaVersion = pana.packageVersion;