+++ CHANGELOG.md
@@ -2,6 +2,11 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2021.06.14`.
+ * Upgraded stable Dart analysis SDK to `2.13.3`.
+ * Upgraded preview Dart analysis SDK to `2.13.3`.
+ * Upgraded stable Flutter analysis SDK to `2.2.2`.
+ * Upgraded preview Flutter analysis SDK to `2.2.2`.
 
 ## `20210610t211000-all`